add frontend setup with React, Vite, and Chakra UI; implement basic routing and create HomePage component

+++ be/controllers/product.controller.js
@@ -70,12 +70,18 @@ export const updateProductController = async (req, res) => {
 
 export const deleteProductController = async (req, res) => {
   const { id } = req.params;
+
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res
+      .status(404)
+      .json({ success: false, message: "Invalid product id" });
+  }
   try {
     await Product.findByIdAndDelete(id);
     res
       .status(200)
       .json({ success: true, message: "Product deleted successfully" });
   } catch (error) {
-    res.status(404).json({ success: false, message: "Product not found" });
+    res.status(500).json({ success: false, message: "Internal server error" });
   }
 };
